Prevent e2e and integration tests from running by default
We have no unit tests for this code.

We have a bunch of integration tests that spin up an etcd and api-server,
and it's good that we have them, but they are unsuitable for running as
part of a pre-commit hook.

This commit disables all non-unit tests by default by marking them with the
`integration` tag. That makes it safe to run `go test ./...` and expect it
to complete successfully (because at the moment we have zero unit tests,
but presumably we will work to fix that).

If you want to run the integration tests, you can run:

    make test-integration

Or if you have the binary requirements already in place, you can run:

    go test -tags=integration ./...

The e2e-tests are protected by the `e2e` tag. You can run them via:

    make test-e2e

(This is unchanged, although we have modified the Makefile to use build
tags rather than `grep` to target specific tests.)